Refactor faucet
